Rename controller to reconciler
For consistency with serving.
