+++ contrib/natss/pkg/controller/channel/reconcile_test.go
@@ -21,7 +21,7 @@ import (
 	"testing"
 
 	eventingv1alpha1 "github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
-	controllertesting "github.com/knative/eventing/pkg/controller/testing"
+	controllertesting "github.com/knative/eventing/pkg/reconciler/testing"
 	"github.com/knative/eventing/pkg/provisioners"
 	util "github.com/knative/eventing/pkg/provisioners"
 	duckv1alpha1 "github.com/knative/pkg/apis/duck/v1alpha1"
