+++ pkg/provisioners/channel_util.go
@@ -16,7 +16,7 @@ import (
 	runtimeClient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	eventingv1alpha1 "github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
-	"github.com/knative/eventing/pkg/controller"
+	"github.com/knative/eventing/pkg/reconciler"
 	"github.com/knative/eventing/pkg/system"
 	"github.com/knative/eventing/pkg/utils"
 	"k8s.io/apimachinery/pkg/api/equality"
@@ -315,7 +315,7 @@ func virtualOldServiceLabels(c *eventingv1alpha1.Channel) map[string]string {
 // appropriate OwnerReferences on the resource so handleObject can discover the Channel resource
 // that 'owns' it. As well as being garbage collected when the Channel is deleted.
 func newVirtualService(channel *eventingv1alpha1.Channel, svc *corev1.Service) *istiov1alpha3.VirtualService {
-	destinationHost := controller.ServiceHostName(channelDispatcherServiceName(channel.Spec.Provisioner.Name), system.Namespace)
+	destinationHost := reconciler.ServiceHostName(channelDispatcherServiceName(channel.Spec.Provisioner.Name), system.Namespace)
 	return &istiov1alpha3.VirtualService{
 		ObjectMeta: metav1.ObjectMeta{
 			GenerateName: channelVirtualServiceName(channel.Name),
@@ -331,7 +331,7 @@ func newVirtualService(channel *eventingv1alpha1.Channel, svc *corev1.Service) *
 		},
 		Spec: istiov1alpha3.VirtualServiceSpec{
 			Hosts: []string{
-				controller.ServiceHostName(svc.Name, channel.Namespace),
+				reconciler.ServiceHostName(svc.Name, channel.Namespace),
 				channelHostName(channel.Name, channel.Namespace),
 			},
 			Http: []istiov1alpha3.HTTPRoute{{
