+++ pkg/provisioners/gcppubsub/dispatcher/dispatcher/reconcile_test.go
@@ -39,7 +39,7 @@ import (
 	"github.com/knative/eventing/pkg/provisioners/gcppubsub/util/fakepubsub"
 
 	eventingv1alpha1 "github.com/knative/eventing/pkg/apis/eventing/v1alpha1"
-	controllertesting "github.com/knative/eventing/pkg/controller/testing"
+	controllertesting "github.com/knative/eventing/pkg/reconciler/testing"
 	"go.uber.org/zap"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
