+++ pkg/reconciler/controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package controller
+package reconciler
 
 import (
 	kubeinformers "k8s.io/client-go/informers"
