+++ pkg/reconciler/eventing/inmemory/controller/kodata/LICENSE
