+++ pkg/reconciler/eventing/inmemory/controller/kodata/VENDOR-LICENSE
